python-ai/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
python-multipart
# ocr
easyocr
pillow
numpy
# validation
pydantic
